Fix bug authen

+++ app/Http/Controllers/ResetPasswordController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Carbon\Carbon;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Str;
 use App\Models\User;
 use Illuminate\Http\Request;
@@ -18,7 +20,7 @@ class ResetPasswordController extends Controller
      * @param Request $request
      * @return JsonResponse
      */
-    public function sendMail(Request $request): JsonResponse
+    public function send_mail_reset_password(Request $request): JsonResponse
     {
         $token = Str::random(60);
         try {
@@ -43,7 +45,35 @@ class ResetPasswordController extends Controller
         ]);
     }
 
-    public function reset(Request $request)
+    public function change_password(Request $request): JsonResponse
+    {
+        $user = $request->user;
+        $passwords = $request->only('old_password', 'new_password');
+        $validator = Validator::make($passwords, [
+            'old_password' => 'required|string|min:6|max:50',
+            'new_password' => 'required|string|min:6|max:50'
+        ]);
+        if ($validator->fails()) {
+            return response()->json([
+                'success' => false,
+                'message' => 'New password is weak (required 6 - 50 characters).',
+            ], 400);
+        }
+        if (!Hash::check($passwords["old_password"], $user->password))
+        {
+            return response()->json([
+                'success' => false,
+                'message' => 'Old password is not correct.',
+            ], 409);
+        }
+        $user->update(["password" => bcrypt($passwords["new_password"])]);
+        return response()->json([
+            'success' => true,
+            'message' => 'Password reset successfully.'
+        ]);
+    }
+
+    public function reset_password(Request $request): JsonResponse
     {
         $token = $request->input("token");
         try {
@@ -62,11 +92,20 @@ class ResetPasswordController extends Controller
             ], 404);
         }
         $user = User::where('email', $passwordReset->email)->firstOrFail();
-        $updatePasswordUser = $user->update(["password" => bcrypt($request->only('password')["password"])]);
+        $new_password = $request->only('password');
+        $validator = Validator::make($new_password, [
+            'password' => 'required|string|min:6|max:50'
+        ]);
+        if ($validator->fails()) {
+            return response()->json([
+                'success' => false,
+                'message' => 'New password is weak (required 6 - 50 characters).',
+            ], 400);
+        }
+        $user->update(["password" => bcrypt($new_password["password"])]);
         $passwordReset->delete();
-
         return response()->json([
-            'success' => $updatePasswordUser,
+            'success' => true,
             'message' => 'Password reset successfully.'
         ]);
     }
